+++ packages/widgetbook/lib/src/routing/app_route_parser.dart
@@ -5,12 +5,29 @@ import 'app_route_config.dart';
 
 @internal
 class AppRouteParser extends RouteInformationParser<AppRouteConfig> {
+  /// This allows a value of type T or T?
+  /// to be treated as a value of type T?.
+  ///
+  /// We use this so that APIs that have become
+  /// non-nullable can still be used with `!` and `?`
+  /// to support older versions of the API as well.
+  T? _ambiguate<T>(T? value) => value;
+
   @override
   Future<AppRouteConfig> parseRouteInformation(
     RouteInformation routeInformation,
   ) async {
+    // Not backwards compatible with Flutter < 3.13.0
+    // ignore: deprecated_member_use
+    final location = routeInformation.location;
+
+    // We use [_ambiguate] to support older versions of Flutter,
+    // since [RouteInformation.location] has changed to be non-nullable
+    // in Flutter 3.13.0.
+    final ambiguousLocation = _ambiguate(location);
+
     return AppRouteConfig(
-      location: routeInformation.location ?? '/',
+      uri: Uri.parse(ambiguousLocation ?? '/'),
     );
   }
 
@@ -19,7 +36,9 @@ class AppRouteParser extends RouteInformationParser<AppRouteConfig> {
     AppRouteConfig configuration,
   ) {
     return RouteInformation(
-      location: configuration.location,
+      // Not backwards compatible with Flutter < 3.13.0
+      // ignore: deprecated_member_use
+      location: configuration.uri.toString(),
     );
   }
 }